+++ src/main/java/blackjack/domain/card/generator/DeckGenerator.java
@@ -1,4 +1,6 @@
-package blackjack.domain;
+package blackjack.domain.card.generator;
+
+import blackjack.domain.card.Card;
 
 import java.util.List;
