+++ src/main/java/blackjack/domain/user/Player.java
@@ -1,11 +1,14 @@
-package blackjack.domain;
+package blackjack.domain.user;
+
+import blackjack.domain.card.Card;
+import blackjack.domain.card.CardGroup;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
 public class Player extends User {
 
-    protected Player(String name, CardGroup cardGroup) {
+    public Player(String name, CardGroup cardGroup) {
         super(name, cardGroup);
     }
