+++ src/main/java/blackjack/domain/user/User.java
@@ -1,4 +1,8 @@
-package blackjack.domain;
+package blackjack.domain.user;
+
+import blackjack.domain.card.Card;
+import blackjack.domain.card.CardGroup;
+import blackjack.domain.card.Deck;
 
 import java.util.List;
