+++ src/main/java/blackjack/domain/user/Users.java
@@ -1,4 +1,9 @@
-package blackjack.domain;
+package blackjack.domain.user;
+
+import blackjack.domain.GameResult;
+import blackjack.domain.card.Card;
+import blackjack.domain.card.CardGroup;
+import blackjack.domain.card.Deck;
 
 import java.util.HashMap;
 import java.util.List;
