+++ src/main/java/blackjack/dto/CardResult.java
@@ -1,6 +1,6 @@
 package blackjack.dto;
 
-import blackjack.domain.Card;
+import blackjack.domain.card.Card;
 
 import java.util.List;
