+++ src/test/java/blackjack/domain/card/DeckTest.java
@@ -1,5 +1,6 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
+import blackjack.domain.card.generator.TestDeckGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
