+++ src/test/java/blackjack/domain/card/generator/RandomDeckGeneratorTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.card.generator;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
