+++ src/test/java/blackjack/domain/user/UsersTest.java
@@ -1,5 +1,12 @@
-package blackjack.domain;
+package blackjack.domain.user;
 
+import blackjack.domain.GameResult;
+import blackjack.domain.card.Card;
+import blackjack.domain.card.CardNumber;
+import blackjack.domain.card.CardShape;
+import blackjack.domain.card.Deck;
+import blackjack.domain.card.generator.RandomDeckGenerator;
+import blackjack.domain.card.generator.TestDeckGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
